(PC-22949) move cheat menu button from home to profile
